+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fix rare failures of `wait_for_state_change` function that resulted in early timeouts.
 * All keywords on nodes are correctly supported, and not only the first one for each option.
 * Function `wait_for_state_change` now supports strings and enums for keywords nodes.
+* Fix #252. ListNodes now also supports numpy integer types as index argument.
 
 ## Version 0.6.0
 * Revert full support of `fnmatch` wildcards and instead use the LabOne wildcard support.